Created request payload util to convert requeset payloads into streams
