+++ src/ds3/models/bulk.go
@@ -1,7 +1,6 @@
 package models
 
 import (
-    "encoding/xml"
     "net/http"
     "ds3/networking"
 )
@@ -22,38 +21,6 @@ type objects struct {
 type bulkObject struct {
     Name string `xml:"name,attr"`
     Size int64 `xml:"size,attr"`
-}
-
-type ds3ObjectList struct {
-    XMLName xml.Name
-    Ds3Objects []Ds3Object `xml:"object"`
-}
-
-func newDs3ObjectList(ds3objects []Ds3Object) *ds3ObjectList {
-    return &ds3ObjectList{
-        XMLName: xml.Name{Local:"objects"},
-        Ds3Objects: ds3objects,
-    }
-}
-
-type Ds3Object struct {
-    Name string `xml:"name,attr"`
-    Size int64 `xml:"size,attr"`
-}
-
-// Converts the ds3 object list to an object we can send in our request.
-func buildDs3ObjectListStream(ds3Objects []Ds3Object) networking.ReaderWithSizeDecorator {
-    // Build the ds3 object list entity.
-    objects := newDs3ObjectList(ds3Objects)
-
-    // Create an xml document from the entity.
-    xmlBytes, err := xml.Marshal(objects)
-    if err != nil {
-        panic(err)
-    }
-
-    // Create a ByteReaderWithSizeDecorator which the network layer expects.
-    return networking.BuildByteReaderWithSizeDecorator(xmlBytes)
 }
 
 // Parses the DS3 specific bulk command response.
